api/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
firebase_admin==6.5.0
google-cloud-firestore==2.16.0
google-cloud-logging==3.10.0
google-cloud-storage==2.17.0
google-cloud-container==2.47.0
google-cloud-workflows==1.14.4
google-api-python-client==2.133.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
requests==2.32.3